pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hypersniper"
version = "0.1.0"
description = "HyperSniper — ultra-fast TON alpha scanner и трейдинг бот"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "HyperSniper Core"}]
dependencies = [
    "aiogram>=3.22.0,<4.0.0",
    "aiohttp>=3.10.0",
    "aiocache>=0.12.2",
    "loguru>=0.7.2",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.1",
    "tonsdk>=1.0.15",
    "pytonconnect @ git+https://github.com/XaBbl4/pytonconnect",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "ruff>=0.7.0"]

[tool.ruff]
line-length = 100
target-version = "py312"

